test: split context and provider and fix land test
Issue #480

Having context and provider splited into two files make it possible to
mock values for testing components.
