+++ frontend/src/__test__/views/Landing.test.tsx
@@ -5,14 +5,27 @@ import { render } from '@testing-library/react';
 import renderer from 'react-test-renderer';
 import Landing from '../../views/Landing';
 import '@testing-library/jest-dom';
-import { describe, it, expect } from 'vitest';
+import { describe, it, expect, vi } from 'vitest';
+import { AuthContext } from '../../contexts/AuthContext';
 
 describe('Landing component test', () => {
+  const contextValue = {
+    signed: true,
+    user: null,
+    isCurrentAuthUser: vi.fn(),
+    signIn: vi.fn(),
+    signOut: vi.fn(),
+    provider: 'idir',
+    token: 'anything'
+  };
+
   it('should have the correct title', () => {
     const { getByTestId } = render(
-      <BrowserRouter>
-        <Landing />
-      </BrowserRouter>
+      <AuthContext.Provider value={contextValue}>
+        <BrowserRouter>
+          <Landing />
+        </BrowserRouter>
+      </AuthContext.Provider>
     );
 
     const content = {
@@ -34,9 +47,11 @@ describe('Landing component test', () => {
   it('should match the snapshot', () => {
     const landing = renderer
       .create(
-        <BrowserRouter>
-          <Landing />
-        </BrowserRouter>
+        <AuthContext.Provider value={contextValue}>
+          <BrowserRouter>
+            <Landing />
+          </BrowserRouter>
+        </AuthContext.Provider>
       ).toJSON();
 
     expect(landing).toMatchSnapshot();
